Added stagger and fixed some sunder typos
