+++ eldritchmush/commands/combat.py
@@ -368,99 +368,6 @@ class CmdResist(Command):
             self.caller.msg("|400You have 0 resists remaining or do not have the skill.")
 
 
-class CmdDisarm(Command):
-    """
-    Issues a disarm command.
-
-    Usage:
-
-    disarm <target>
-
-    This will issue a disarm command that reduces the next amount of damage taken by master of arms level.
-    """
-
-    key = "disarm"
-    help_category = "mush"
-
-    def parse(self):
-        "Very trivial parser"
-        self.target = self.args.strip()
-
-    def func(self):
-            h = Helper()
-
-            "Get level of master of arms for base die roll. Levels of gear give a flat bonus of +1/+2/+3."
-            disarmsRemaining = self.caller.db.disarm
-            master_of_arms = self.caller.db.master_of_arms
-            hasBow = self.caller.db.bow
-            hasMelee = self.caller.db.melee
-            weapon_level = h.weaponValue(self.caller.db.weapon_level)
-            hasTwoHanded = self.caller.db.twohanded
-
-            wylding_hand = self.caller.db.wyldinghand
-
-            if not self.args:
-                self.caller.msg("|540Usage: disarm <target>|n")
-                return
-
-            target = self.caller.search(self.target)
-
-            if not target:
-                self.caller.msg("|400There is nothing here that matches that description.|n")
-                return
-
-            if target == self.caller:
-                self.caller.msg(f"|400Don't disarm yourself {self.caller}!|n")
-                return
-
-            if target.db.body is None:
-                self.caller.msg("|400You had better not try that.")
-                return
-
-            if target.db.bow:
-                self.caller.msg("|400This attack may not be used on archers.|n")
-
-            # Check for weakness on character
-            weakness = h.weaknessChecker(self.caller.db.weakness)
-
-            # Check for equip proper weapon type or weakness
-            if weakness:
-                self.caller.msg("|400You are too weak to use this attack.|n")
-            elif hasBow:
-                self.caller.msg("|540Before you can attack, you must first unequip your bow using the command setbow 0.")
-            elif not hasMelee or not hasTwoHanded:
-                self.caller.msg("|540Before you can attack, you must first equip a melee weapon using the command setmelee 1 or settwohanded 1.")
-            else:
-                if disarmsRemaining > 0:
-                # Return die roll based on level in master of arms or wylding hand.
-                    if wylding_hand:
-                        die_result = h.wyldingHand(wylding_hand)
-                    else:
-                        die_result = h.masterOfArms(master_of_arms)
-
-                    # Decrement amount of disarms from amount in database
-                    self.caller.db.disarm -= 1
-
-                    # Get final attack result and damage
-                    weakness = h.weaknessChecker(self.caller.db.weakness)
-                    dmg_penalty = h.bodyChecker(self.caller.db.body)
-                    attack_result = (die_result + weapon_level) - dmg_penalty - weakness
-
-                    # Return attack result message
-                    if attack_result >= target.db.av:
-                        self.caller.location.msg_contents(f"|015{self.caller.key} skillfully disarms {target.key}.|n")
-                        target.msg("|540You have been disarmed. Equip your weapon by using the commands, setmelee 1, settwohanded 1, or setbow 1.|n\n")
-                        if target.db.melee:
-                            target.db.melee = 0
-                        elif target.db.twohanded:
-                            target.db.twohanded = 0
-                        self.caller.msg(f"|540You disarm {target.key}|n")
-                    elif attack_result < target.db.av:
-                        self.caller.location.msg_contents(f"|015{self.caller.key} attempts|n (|400{attack_result}|n)|015 to disarm {target.key}|n (|020{target.db.av}|n)|015, but fumbles their attack.|n")
-                else:
-                    self.caller.msg("|400You have 0 disarms remaining or do not have the skill.")
-
-
 class CmdStun(Command):
     """
     Issues a stun command.
